Remove ember-simple-auth and add custom auth service
ember-simple-auth requires the identification parameter (usually email)
that is used for log in to be the same that is also sent next to the token
in the Authorization header. Since obsidian-web requires `name` to be
used for log in and `secret_id` to be sent in the header, I decided to
implement my own auth service (service:auth) which is automatically injected
into the routes and initialised upon startup. It uses HTML5's
window.sessionStorage to store school_id, secret_id, and the token. The
method authenticate(name, password) initiates an ajax request to the backend's
login endpoint and upon successful login stores the returned token, secret_id,
and school_id. invalidate() sends a request to the logout endpoint and deletes
the session data. init(), which is called upon startup, registers an ajax
prefilter that automatically adds the Authorization header when it it required.