Add models and customise ApplicationAdapter
Depending on the current environment, a different host is being set for
ember-data. Apart from that, the paths are now generated by pluralising and
underscoring the model name (instead of pluralising and dasherising it).

No model belongs to a school because all request are automatically scoped on
server side. The school model merely exists to perform account settings and
to add new schools.