Add templates and controllers for login, logout, and register.
Clicking on `logout` invokes the application controller's logout action which
redirects upon successful logout to the index page. `Login` initiates a
transition to the /login route, which displays a login form. Submitting the form
invokes the login controller's login action which uses the auth-service to
log in and redirect when successful. Register does essentially the same except
for creating and persisting a school record before. Everything is chained nicely
with promises and the then() function.