Remove custom serializer; Fix lendings-(list | entry)'s date display
In correlation with 9a584fba071e49a776307ab6eb62709f31a6645a@obsidian-web,
I removed the custom serializer. When showing a student's lendings, an error
used to be thrown ('this' not defined), which was due to the fact that the
date (as Ember.computed) was defined using an ES6 lambda () => {} which doesn't
initialise the context. However, function() {} does and it works smoothly.
As the date is now properly rendered, it became obvious that the date column
needed to be wider.