Pull common functionality of lend-base and return-base controllers into common superclass.
Lend-base's and return-base's controllers (as well as lend's and return's)
all contain (or will contain) a lookup method for books when their isbn or alias
is entered and their save() actions all rely on looping through all isbns and
doing something (the something is implemented by each controller using the
processIsbn hook) and then transitioning (by calling the transition action).
The controllers only have to implement the processIsbn hook and the redirect
action.