Change LendingsController (common superclass of all controllers which handle lendings)
I changed redirect() to be a method instead of an action, and most importantly,
I modified the save() action so that the "this" context is now automatically
injected into every call to processIsbn() ("Unknown attribute 'store' of
undefined" used to be thrown). Furthermore, I made the students.return-base
controller extend LendingsController.