Fix "delegate is not a function" error on students.index
The main cause for the error is, that I appearently did not set a default
value for the klass property in the class definition. So I moved the assignment
of the default value from the constructor to the class definition. Moreover,
I renamed the parameter passed to class-list (from 'klasses' to 'data', like
in the other *-lists) and I removed the superfluous property 'klass' from
class-entry and replaced all its occurrences by 'value'.