Add list view
I added some styling to app.scss to support rotated table headings. Apart from
that, I added two query params to students.index: style and list. Style can
take the values 'table' and 'list' and determines whether a simple
representation of the student *table* or a more sophisticated *list* of students
(of a specific class) and their baseSets shall be rendered. The table has been
implemented before, and the list is handled by the student-table component
which yields whether a the current column's book has been lent by the current
row's student as a baseSet. Since I already anticipated the style param (though
under a different name, which seems to be reserved in templates: view), I needed
to change its name.
Proper reloading of all data remains to be implemented because upon testing, the
page tried to load the book of a baseSet, which has just been deleted.