Fix editing student's class
Student.set('klass') used to return undefined which made the app think
something went wrong, which led me to believe, computedProperties shouldn't
be bound to input fields. So I used an intermediate, the property klass of
the student-entry component. The problem is, I didn't update the input field
to set the intermediate. That way, it always reset the student's class upon
saving.