Move student creation logic to students.index's controller
Instead of using the store that used to be passed as param,
import-dialog now calles an action on students.index's
controller which creates and saves the student record.
import-dialog#submit itself only implements the parsing of
headers and the input file and assembles the acquired data
in records it then passes to its action.