Add tiles layout to /index template
I added the masonry jQuery plugin. Furthermore, I added jQuery invocations to
adjust the tiles' heights and to colour them in random shades of blue.
Furthermore, I also added those jQuery method calls to the didTransition hook
of /index's route (so that when the user transitions to the page, the layout
is still being jQuery'd).