Fix error occurring when adding a new book and studentTable's display
Turns out, the reason for the page's not navigating back to book's index view
after adding a new book is that I misspelt flashMessages which threw an error
that got subsequently handled by the promise's error handler (which attempted
to print an error message via the same method and failed too). So in the end,
nothing has been done after adding a new book. But I fixed the typo and it works
now.

StudentTable used to truncate the name of the last column (the size estimates
are not influenced by the rotation), so I added a ghost field to the header that
takes the necessary space, so that the "last" header column is now properly
visible