Fix further layout issues
In student-table, the buttons are finally displayed next to each other (because
the promise that loads the data also causes a fake resizing event that
adjusts the respective widths). Furthermore, the scrolling is now mostly
constricted to a div that wraps the table (this way, one does not scroll "past"
the navbar).